requirements.txt
libclang
typer
rich
jinja2
pyyaml
google-generativeai
openai
python-dotenv
requests
